Update to support replying to a Tweet
Why:
Using the reply form was not allowing for the update of replies for any given
Tweet. This was not allowing the new created tweet from being added to
the replies array within a Tweet object.

Updated Tweet was not updating in the list view.

How:
Added additinonal functionality to the Form to handle the submission of Tweets
appropriately. Submitting the Tweet first finds the targeted object Tweet and
push's the new created Tweet object into the array of replies. (To the end of
the list.) UUID was installed to handle dynamic unique ID generation.

The detail screen hook was modified to support pulling Tweets from the array of
Tweets in the Global State context. This allows for instant update to the list view
within the Detail Screen component.

Tags:
refactor, bug-fix, uuid